Move course script and fix test
